Made a custom Eshell prompt that I'm satisfied with
The prompt is split over two lines (so long folder names are not an issue),
and only shows *necessary* information. In other words, it shows the full
TRAMP path (with proper colorization); but it hides the hostname when I'm
on localhost, hides the folder when I'm in ~/, and always hides the username
as I usually know who I am (one username per machine, never log in as root).

I also removed the Ivy/Counsel configuration, as I've switched to Helm and
am quite happy with that for now. There's also not much reason to keep C-RET
as a Helm keybinding, since C-SPC is actually reasonable with Caps/Ctrl remap.

As a minor note, I remembered to turn off centered-point in shell buffers.