Updated LSP/Flycheck settings
Now, `lsp-ui` is combined with `flake8` in Python buffers, to get
linting in addition to checking. Moreover, the `lsp-ui` frontend
has gotten a number of built-in features disabled.

As minor other changes, I made `vterm` use `fish` even if the system
is set to use `bash` (for POSIX-compatibility), and got rid of the
`vi-tilde-fringes` in a way that doesn't cause errors on buffer opens.