Clean: Removed most of my remaining NeoVim config.
I rarely use NeoVim for more than quick edits after switching to Emacs,
so most of my config could be stripped away. However, I wanted to add
the Moonlight theme for consistency with Emacs when I do use it, as well
as a plugin for editing Org files in case I need terminal access to that.