Feat: Alacritty + Tmux setup.
Kitty is very nice, but I don't have access to it on Windows/WSL
or over SSH. The benefit of using Tmux is then that I can have
same kind of tabbing on every platform, locally and remotely.

This is a basic setup that looks like my previous Kitty setup,
just implementing the "tabs" in Tmux, with a statusline that
integrates with Ubuntu Yaru above and Solarized Light below.